third_party/blink/web_tests/external/wpt/css/css-conditional/container-queries/scroll-state-snapped-wm-expected.txt

This is a testharness.js-based test.
[FAIL] Check scroll-state(snapped) matching dynamically changing to horizontal writing-mode
  assert_equals: expected "block" but got "inline"
Harness: the test ran to completion.
